% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{convert_percentages_to_counts}
\alias{convert_percentages_to_counts}
\title{Convert percentage columns to counts}
\usage{
convert_percentages_to_counts(df)
}
\arguments{
\item{df}{Data frame with potential percentage columns}
}
\value{
Data frame with counts
}
\description{
Some ALSDE data exports provide percentages rather than counts.
This function converts them using the row_total.
}
\keyword{internal}
