% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_al_id}
\alias{create_al_id}
\title{Create combined Alabama ID}
\usage{
create_al_id(system_code, school_code)
}
\arguments{
\item{system_code}{System (district) code}

\item{school_code}{School code}
}
\value{
Character vector with combined IDs (e.g., "001-0010")
}
\description{
Combines system code and school code into a single identifier
}
\keyword{internal}
