% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{determine_row_type}
\alias{determine_row_type}
\title{Determine row type based on IDs}
\usage{
determine_row_type(df)
}
\arguments{
\item{df}{Data frame with district_id and campus_id columns}
}
\value{
Character vector of types
}
\description{
Determine row type based on IDs
}
\keyword{internal}
