% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_acap_assessment}
\alias{download_acap_assessment}
\title{Download ACAP Reading assessment data}
\usage{
download_acap_assessment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with assessment data
}
\description{
The Alabama Achieves website provides ACAP Reading assessment data in Excel format.
This function downloads the appropriate file for the requested school year.
}
\details{
File URLs follow this pattern:
\itemize{
\item 2021-2022: https://www.alabamaachieves.org/wp-content/uploads/2022/07/2021-2022-ACAP-Reading.xlsx
\item 2022-2023: https://www.alabamaachieves.org/wp-content/uploads/2023/06/RD_SP_202368_2022-2023-ACAP-Reading_V1.0.xlsx
\item 2023-2024: https://www.alabamaachieves.org/wp-content/uploads/2024/06/RD_SP_2024613_2023-2024-ACAP-Reading.xlsx
\item 2024-2025: https://www.alabamaachieves.org/wp-content/uploads/2025/06/RD_SP_2025612_ACAP-Reading_v1.0.xlsx
}
}
\keyword{internal}
