% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_federal_reportcard}
\alias{download_federal_reportcard}
\title{Download data from Federal Report Card Student Demographics}
\usage{
download_federal_reportcard(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with enrollment data
}
\description{
The Federal Report Card provides detailed enrollment data by demographics
at the school level. This function downloads the CSV export.
}
\details{
The Student Demographics page is available at:
https://reportcard.alsde.edu/SupportingData_StudentDemographics.aspx
}
\keyword{internal}
