% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{ensure_aggregation_levels}
\alias{ensure_aggregation_levels}
\title{Ensure all aggregation levels exist}
\usage{
ensure_aggregation_levels(df, end_year)
}
\arguments{
\item{df}{Processed data frame}

\item{end_year}{School year end}
}
\value{
Data frame with all aggregation levels
}
\description{
Creates district and state aggregates if they don't exist in the data.
}
\keyword{internal}
