% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assess}
\alias{get_raw_assess}
\title{Download raw assessment data from ALSDE}
\usage{
get_raw_assess(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024). Valid values are 2022-2025.}
}
\value{
Data frame with raw assessment data
}
\description{
Downloads school-level ACAP Reading assessment data from the Alabama State
Department of Education Alabama Achieves website.
}
\keyword{internal}
