% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw public school directory data from ALSDE}
\usage{
get_raw_directory()
}
\value{
Data frame with raw school directory data
}
\description{
Downloads school-level directory data from the Alabama State Department of
Education Education Directory.
}
\keyword{internal}
