% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment-documentation.R
\name{id_assess_aggs}
\alias{id_assess_aggs}
\title{Add district and state aggregates to assessment data}
\usage{
id_assess_aggs(df)
}
\arguments{
\item{df}{Tidy assessment data frame}
}
\value{
Data frame with added district and state aggregates
}
\description{
Calculates district-level and state-level aggregates from school-level data.
This ensures that totals match the sum of constituent schools.
}
\keyword{internal}
