% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assess}
\alias{process_assess}
\title{Process raw ACAP assessment data}
\usage{
process_assess(raw_data, end_year)
}
\arguments{
\item{raw_data}{Raw data frame from get_raw_assess()}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Standardizes column names and data types from raw ACAP assessment download.
}
\keyword{internal}
