% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_federal_reportcard}
\alias{process_federal_reportcard}
\title{Process Federal Report Card data}
\usage{
process_federal_reportcard(df, end_year)
}
\arguments{
\item{df}{Raw data frame from Federal Report Card}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Process Federal Report Card data
}
\keyword{internal}
